Fix dtype mismatch in depth tensor creation: explicitly use float32
Root cause: torch.as_tensor(depth) was preserving numpy's default float64
dtype from loaded .npy files, causing mismatch with float32 poses/intrinsics
in backprojection (line 128 of backproject.py).

This is the actual source of "RuntimeError: expected scalar type Double but
found Float" - not quat_to_rotmat which was already fixed in 355c7b1.

+++ my_train_odin.py
@@ -501,8 +501,8 @@ class StrawberryDatasetMapper:
             is_white = (rr > 220) & (gg > 220) & (bb > 220)
             is_black = (rr < 20) & (gg < 20) & (bb < 20)
             depth[is_white | is_black] = 0
-            
-            depths.append(torch.as_tensor(depth))
+
+            depths.append(torch.as_tensor(depth, dtype=torch.float32))
             
             # Intrinsics Scaling
             K = intrinsics_list[idx].copy()
